building test

+++ dhl-api/src/main/scala/com/mailytica/ai/api/freshdesk/FreshdeskMail.scala
@@ -32,7 +32,7 @@ object FreshdeskMail {
 
     Some(FreshdeskMail(
       FreshdeskBodyText(freshdeskBodyText),
-      FreshdeskId(freshdeskId),
+      FreshdeskId(freshdeskId.toInt),
       FreshdeskSupportEmail(freshdeskSupportEmail),
       FreshdeskToEmails(freshdeskEmailSeq)
     ))
